Created Weather API City search bar
Removed the non functioning search bar and added info from the open weather api. Enter a city and pulls data of current weather. Please note in the plants.py there is a API key required. Its currently as '****' you will need to add a API key for this to work. Basic as no styling added, but thoughts?